+++ app.trading.algoritmico.api/src/AppTradingAlgoritmico.Domain/Enums/PlatformType.cs
@@ -1,0 +1,7 @@
+namespace AppTradingAlgoritmico.Domain.Enums;
+
+public enum PlatformType
+{
+    MT4 = 0,
+    MT5 = 1
+}