+++ app.trading.algoritmico.api/src/AppTradingAlgoritmico.Infrastructure/Persistence/AppDbContext.cs
@@ -8,6 +8,8 @@ public class AppDbContext : IdentityDbContext<ApplicationUser, ApplicationRole, 
 {
     public AppDbContext(DbContextOptions<AppDbContext> options) : base(options) { }
 
+    public DbSet<TradingAccount> TradingAccounts => Set<TradingAccount>();
+
     protected override void OnModelCreating(ModelBuilder builder)
     {
         base.OnModelCreating(builder);